feat: split library
